Move the message queues into endpoint.py
