--- gpgsync/message_queues.py
@@ -1,48 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-GPG Sync
-Helps users have up-to-date public keys for everyone in their organization
-https://github.com/firstlookmedia/gpgsync
-Copyright (C) 2016 First Look Media
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
-"""
-import queue
-
-
-class VerifierMessageQueue(queue.Queue):
-    def __init(self):
-        super(VerifierMessageQueue, self).__init__()
-
-    def add_message(self, msg=None, type='update', timeout=0):
-        self.put({
-            'type': type,
-            'msg': msg,
-            'timeout': timeout
-        })
-
-
-class RefresherMessageQueue(queue.LifoQueue):
-    STATUS_STARTING = 0
-    STATUS_IN_PROGRESS = 1
-
-    def __init(self):
-        super(RefresherMessageQueue, self).__init__()
-
-    def add_message(self, status, total_keys=0, current_key=0):
-        self.put({
-            'status': status,
-            'total_keys': total_keys,
-            'current_key': current_key
-        })
